etl_client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "animal-etl-client"
version = "0.1.0"
description = "ETL client for processing animal data from the Animals API"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "requests==2.31.0",
    "pydantic==2.11.9",
    "python-dotenv==1.0.0",
    "click==8.1.7",
    "tenacity==8.2.3",
    "structlog==23.2.0",
    "colorama==0.4.6"
]
requires-python = ">=3.8"
readme = "README.md"

[project.scripts]
animal-etl = "etl_client.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["etl_client*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tqdm.*",
]
ignore_missing_imports = true
